requirements.txt
# Core dependencies
pyspark==3.5.0
pyyaml==6.0.1
pydantic==2.5.0
pandas==2.1.4
pyarrow==14.0.1

# Cloud connectors
boto3==1.34.0              # AWS S3
azure-storage-file-datalake==12.14.0  # Azure ADLS Gen2
google-cloud-storage==2.14.0  # Google Cloud Storage

# Database connectors (optional)
pyhive==0.7.0              # Hive
thrift==0.20.0
thrift-sasl==0.4.3
sasl==0.3.1
psycopg2-binary==2.9.9     # PostgreSQL
mysql-connector-python==8.2.0  # MySQL

# Reporting
jinja2==3.1.2
plotly==5.18.0
rich==13.7.0

# CLI and utilities
click==8.1.7
python-dotenv==1.0.0
tenacity==8.2.3            # Retry logic

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# Development tools
black==23.12.1
flake8==7.0.0
mypy==1.7.1
